Update index.js to correctly log Insight kWh
The logged value of kWh is being calculated incorrectly.  This became evident to me as I had an Insight switch pulling 22mW, however, >20kWh was being reported for daily use (20kWh is a device pulling 1kW for 20 hours).  Problem was equation in line 834.  'raw' in this equation is data.TodayConsumed (line 56) which is in 'mW minutes'.  To convert to kWh you need to divide 'raw" by 60 to convert to hours and then by 1,000,000 to convert from 'mW hours' to 'kWh'.  As this would have resulted in a value that was almost always zero for a low power draw I calculated 'value' as 'Wh' and added a separate equation for 'kWh'.  In the process of debugging I also included 'TodayONTime' converted to hours as a log entry.  A list of changes is as follows:
- [ ] line 56 - change unit from ‘kWh’ to ‘Wh’ (as raw is UINT32 and low mW draw would truncate to zero if direct convert to kWh)
- line 665 - add second parameter “data.TodayONTime”
- line 833 - add “raw2” parameter from line 665 change
- line 834 - correct equation to convert mW minutes to Wh
- new line 835 to calculate kWh
- new line 836 to calculate onHours
- add line 841 to log ‘onHours’
- line 842 - updated to log ‘kWh’